Create new RunnableSerializable class in preparation for configurable runnables
- Also move RunnableBranch to its own file
