+++ libs/langchain/langchain/schema/runnable/passthrough.py
@@ -16,9 +16,13 @@ from typing import (
     cast,
 )
 
-from langchain.load.serializable import Serializable
 from langchain.pydantic_v1 import BaseModel, create_model
-from langchain.schema.runnable.base import Input, Runnable, RunnableMap
+from langchain.schema.runnable.base import (
+    Input,
+    Runnable,
+    RunnableMap,
+    RunnableSerializable,
+)
 from langchain.schema.runnable.config import RunnableConfig, get_executor_for_config
 from langchain.schema.runnable.utils import AddableDict
 from langchain.utils.aiter import atee, py_anext
@@ -33,7 +37,7 @@ async def aidentity(x: Input) -> Input:
     return x
 
 
-class RunnablePassthrough(Serializable, Runnable[Input, Input]):
+class RunnablePassthrough(RunnableSerializable[Input, Input]):
     """
     A runnable that passes through the input.
     """
@@ -109,7 +113,7 @@ class RunnablePassthrough(Serializable, Runnable[Input, Input]):
             yield chunk
 
 
-class RunnableAssign(Serializable, Runnable[Dict[str, Any], Dict[str, Any]]):
+class RunnableAssign(RunnableSerializable[Dict[str, Any], Dict[str, Any]]):
     """
     A runnable that assigns key-value pairs to Dict[str, Any] inputs.
     """
